Type hotfix.

+++ src/core/ngx_connection.c
@@ -365,7 +365,7 @@ ngx_set_inherited_sockets(ngx_cycle_t *cycle)
             continue;
         }
 
-        if (olen < sizeof(int) || timeout == 0) {
+        if (olen < (int)sizeof(int) || timeout == 0) {
             continue;
         }
